Use jsoniter unmarshaller as default trace unmarshaler. Update unit test style design.
